+++ NEWS.md
@@ -1,3 +1,8 @@
+# caugi 0.3.1
+
+* Remove the use of `lockBinding` and `unlockBinding` in the package to 
+  silence R CMD check notes.
+
 # caugi 0.3.0
 
 * Add `mutate_caugi` function that allows conversion from one graph type to another.